requirements/prod.txt
Django==2.2.2
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
djangorestframework-camel-case==1.0.3
six==1.12.0
coreapi==2.3.3
psycopg2-binary==2.7.5
gunicorn==19.9.0


